requirements.txt
keyboard==0.13.5
mss==7.0.1
numpy==1.21.6
Pillow==9.5.0
Pillow==10.0.0
PyAutoGUI==0.9.54
